pyproject.toml
[project]
name = "a2a-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.8",
    "dotenv>=0.9.9",
    "pydantic>=2.11.7",
    "starlette-context>=0.4.0",
    "uvicorn>=0.34.3",
]
